# api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Movie Hub
  description: "A Movie Hub API."
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      description: "The page number for pagination."
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 10000
        default: 0
        example: 0
    limit:
      name: limit
      in: query
      description: "The maximum number of items to return per page."
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 10
        example: 10
...
    companyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
        maxLength: 255
      description: "The production company ID."
      example: 8e9334d2-e021-4ec3-b34b-8dc57032e362
    name:
      name: name
      in: query
      description: "The name of the entity."
      schema:
        type: string
        pattern: "^[A-Za-zÀ-ÿ0-9.,'\"&()\\-: ]*$"
        maxLength: 255
        default: ""
      example: "James Wan"
  responses:
...
    404ProductionCompanyNotFound:
      description: Production company not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    401Unauthorized:
      description: Either no access token was provided or the provided access token is invalid or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    403Forbidden:
      description: The user does not have permission to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
    Uuid:
      type: string
      format: uuid
      description: |
        A universally unique identifier (UUID) used to uniquely identify entities.
        It is a 36-character string consisting of hexadecimal digits and hyphens.
      example: 872e1723-09ac-461d-9f54-d4c221269bed
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
      maxLength: 255
      nullable: false
    MovieName:
      type: string
      description: |
        The title of a movie.
        This field should contain the movie's name and is restricted to a maximum length of 100 characters.
      example: The Conjuring
      pattern: "^[A-Za-z0-9.,'\"&()\\-: ]+$"
      minLength: 1
      maxLength: 100
      nullable: false
...
    MovieReleaseYear:
      type: integer
      format: int32
      description: |
        The release year of the movie.
        This indicates when the movie was officially released.
      example: 2024
      nullable: false
    MovieDuration:
      type: integer
      format: int32
      description: |
        The duration of the movie in minutes.
        This field must be a positive integer representing the total running time of the movie.
      example: 112
      minimum: 1
      maximum: 1000
      nullable: false
...
    MoviePosterUrl:
      type: string
      description: |
        The URL to the movie's poster image.
        This link points to the visual representation of the movie used for promotional purposes.
      example: 'https://image.tmdb.org/t/p/original/wVYREutTvI2tmxr6ujrHT704wGF.jpg'
      pattern: "^(https?://)?(www\\.)?[-a-zA-Z0-9@:%._+~#=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_+.~#?&//=]*)$"
      minLength: 1
      maxLength: 2000
      nullable: false
...
    ProductionCompanyName:
      type: string
      example: A24
      description: |
        The name of the company that produced the movie.
        This field contains up to 100 characters and represents the organization responsible for the film's production.
      pattern: "^[\x20-\x7E]+$"
      minLength: 1
      maxLength: 100
      nullable: false
...
    GenreName:
      type: string
      example: Horror
      description: |
        The genre or category of the movie.
        This field represents the type of movie, such as Horror or Comedy, and is limited to 15 characters.
      pattern: "^[A-Za-z\\- ]+$"
      minLength: 1
      maxLength: 15
      nullable: false
...
    SortDTO:
      type: object
      description: "Represents sorting information for ordering query results based on one or more criteria."
      additionalProperties: false
      required:
        - empty
        - sorted
        - unsorted
      properties:
        empty:
          type: boolean
          example: false
          description: |
            A boolean value indicating whether the sorting information is empty. 
            If `true`, there are no sorting criteria defined for the query.
        sorted:
          type: boolean
          example: true
          description: |
            A boolean value indicating whether the query results are sorted. 
            If `true`, the query includes sorting criteria.
        unsorted:
          type: boolean
          example: false
          description: |
            A boolean value indicating whether the query results are unsorted. 
            If `true`, the query results are not sorted according to any criteria.
    PageableDTO:
      type: object
      description: "Represents pagination and sorting information for querying a paginated dataset."
      additionalProperties: false
      required:
        - pageNumber
        - pageSize
        - sort
        - offset
        - paged
        - unpaged
      properties:
        pageNumber:
          type: integer
          description: "The page number to retrieve."
          example: 0
        pageSize:
          type: integer
          description: "The maximum number of elements to be returned on each page."
          example: 2
        sort:
          $ref: '#/components/schemas/SortDTO'
        offset:
          type: integer
          format: int64
          description: "The offset of the current page within the result set."
          example: 0
        paged:
          type: boolean
          description: |
            A boolean value indicating whether the query results are paginated. 
            If `true`, the query includes pagination criteria.
          example: true
        unpaged:
          type: boolean
          description: |
            A boolean value indicating whether the query results are not paginated. 
            If `true`, the query results are not subject to pagination.
          example: false
    MoviePage:
      type: object
      description: "Represents a page of movies, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of movies."
          maxItems: 50
          items:
            $ref: '#/components/schemas/MoviePreviewResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
...
    ProductionCompanyPage:
      type: object
      description: "Represents a page of production companies, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of production companies."
          maxItems: 50
          items:
            $ref: '#/components/schemas/ProductionCompanyDetailsResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
...
    MoviePreviewResponse:
      type: object
      description: "Offers preview representation of the movie."
      additionalProperties: false
      required:
        - id
        - name
        - releaseYear
        - duration
        - posterUrl
        - genres
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/MovieName'
        releaseYear:
          $ref: '#/components/schemas/MovieReleaseYear'
        duration:
          $ref: '#/components/schemas/MovieDuration'
        posterUrl:
          $ref: '#/components/schemas/MoviePosterUrl'
        genres:
          type: array
          maxItems: 50
          items:
            $ref: '#/components/schemas/GenreName'
...
    ProductionCompanyDetailsResponse:
      type: object
      description: "Provides details about a production company."
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/ProductionCompanyName'
...
    ProblemDetail:
      type: object
      title: Problem Detail
      description: "A generic error."
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          description: "A URI reference that identifies the problem type."
          example: "http://localhost:8080/movie-not-found"
          pattern: "^((about:blank)|(https?://[^\\s]+$))"
          maxLength: 1000
        title:
          type: string
          description: "A short, human-readable summary of the problem type."
          example: "Document Not Found"
          pattern: "^[A-Za-z0-9\\s\\-.,!?:;'\"]+$"
          maxLength: 255
        status:
          type: number
          description: "The HTTP status code."
          example: 404
        detail:
          type: string
          description: "A human-readable explanation specific to this occurrence of the problem."
          example: "Movie with ID: 11ae56ed-240e-4d4d-8d3c-f135d9394ab3 not found"
          pattern: "^[A-Za-z0-9\\s\\-.,!?:;'\"]+$"
          maxLength: 1000
        instance:
          type: string
          description: "A URI reference that identifies the specific occurrence of the problem."
          example: "/movies/11ae56ed-240e-4d4d-8d3c-f135d9394ab3"
          pattern: "^(/[^\\s]*)+$"
          maxLength: 1000
        contextInfo:
          type: object
          description: "A free-form based object that contains additional context about the error."
paths:
...
  /production-companies:
    post:
      description: |
        Create a new production company record in the database with the specified details.
        This operation adds a production company to the system, enabling it to be used in association with movies.
      summary: Add production company
      operationId: addProductionCompany
      security:
        - bearerAuth: []
      tags:
        - production-companies
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/ProductionCompanyName'
      responses:
        '201':
          description: Successfully created a new production company record in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionCompanyDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          description: The production company already exists in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: Retrieve a list of production companies available in the system.
      summary: List production companies
      operationId: getProductionCompanies
      tags:
        - production-companies
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/name'
      responses:
        '200':
          description: Successfully retrieved the list of production companies available in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionCompanyPage'
        'default':
          $ref: '#/components/responses/Default'
  /production-companies/{companyId}:
    get:
      description: Retrieve detailed information about a specific production company.
      summary: Get production company details
      operationId: getProductionCompany
      tags:
        - production-companies
      parameters:
        - $ref: '#/components/parameters/companyId'
      responses:
        '200':
          description: Successfully retrieved the details of the specified production company.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionCompanyDetailsResponse'
        '404':
          $ref: '#/components/responses/404ProductionCompanyNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing production company.
        All production company's details will be replaced with the new details provided.
      summary: Update production company details
      operationId: updateProductionCompany
      security:
        - bearerAuth: []
      tags:
        - production-companies
      parameters:
        - $ref: '#/components/parameters/companyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/ProductionCompanyName'
      responses:
        '200':
          description: Successfully updated the details of the specified production company.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductionCompanyDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404ProductionCompanyNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /production-companies/{companyId}/movies:
    get:
      description: Retrieve a list of movies that were produced by the specified company.
      summary: List movies produced by production company
      operationId: getMoviesWithProductionCompany
      tags:
        - production-companies
      parameters:
        - $ref: '#/components/parameters/companyId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Successfully retrieved a list of movies that were produced by the specified company.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        '404':
          $ref: '#/components/responses/404ProductionCompanyNotFound'
        'default':
          $ref: '#/components/responses/Default'
...